# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Farmer Market Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /prices:
    get:
      operationId: listPrices
      tags: [prices]
      summary: List mandi prices
      parameters:
        - name: crop
          in: query
          schema:
            type: string
        - name: market
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of prices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MandiPrice"
  /prices/trends:
    get:
      operationId: getPriceTrends
      tags: [prices]
      summary: Get price trends for a crop
      parameters:
        - name: crop
          in: query
          required: true
          schema:
            type: string
        - name: market
          in: query
          schema:
            type: string
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Price trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceTrend"
  /prices/compare:
    get:
      operationId: comparePrices
      tags: [prices]
      summary: Compare prices across markets
      parameters:
        - name: crop
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Price comparison across markets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketComparison"
...
components:
  schemas:
...
    MandiPrice:
      type: object
      properties:
        id:
          type: integer
        crop:
          type: string
        variety:
          type: string
        market:
          type: string
        state:
          type: string
        district:
          type: string
        minPrice:
          type: number
        maxPrice:
          type: number
        modalPrice:
          type: number
        arrivals:
          type: ["number", "null"]
        priceDate:
          type: string
        createdAt:
          type: string
      required:
        - id
        - crop
        - variety
        - market
        - state
        - district
        - minPrice
        - maxPrice
        - modalPrice
        - arrivals
        - priceDate
        - createdAt
    PriceTrend:
      type: object
      properties:
        date:
          type: string
        modalPrice:
          type: number
        minPrice:
          type: number
        maxPrice:
          type: number
        arrivals:
          type: ["number", "null"]
      required:
        - date
        - modalPrice
        - minPrice
        - maxPrice
        - arrivals
    MarketComparison:
      type: object
      properties:
        market:
          type: string
        state:
          type: string
        modalPrice:
          type: number
        deviation:
          type: number
        rank:
          type: integer
      required:
        - market
        - state
        - modalPrice
        - deviation
        - rank
...